+++ AsBuiltReport.Diagram/Src/Private/ConvertTo-Pdf-WaterMark.ps1
@@ -50,17 +50,17 @@ function ConvertTo-Pdf-WaterMark {
                 'Core' {
                     # Net 9.0 assembly call
                     Write-Verbose -Message 'Successfully loaded the .Net 9.0 assembly for PDF conversion.'
-                    New-ConvertPngToPdf -SourcePngFile $ImageInput.FullName -OutputFileNamePath $DestinationPath
+                    ConvertTo-PngToPdf -SourcePngFile $ImageInput.FullName -OutputFileNamePath $DestinationPath
                 }
                 'Desktop' {
                     Write-Verbose -Message 'Successfully loaded the .Net 4.8 assembly for PDF conversion.'
                     # Net 4.8 assembly call
-                    New-ConvertPngToPdf -SourcePngFile $ImageInput.FullName -OutputFileNamePath $DestinationPath
+                    ConvertTo-PngToPdf -SourcePngFile $ImageInput.FullName -OutputFileNamePath $DestinationPath
                 }
                 default {
                     Write-Verbose -Message 'Successfully loaded the .Net 4.8 assembly for PDF conversion.'
                     # Net 4.8 assembly call (Fucking shit)
-                    New-ConvertPngToPdf -SourcePngFile $ImageInput.FullName -OutputFileNamePath $DestinationPath
+                    ConvertTo-PngToPdf -SourcePngFile $ImageInput.FullName -OutputFileNamePath $DestinationPath
                 }
             }