+++ AsBuiltReport.Diagram/Src/Private/ConvertTo-Svg.ps1
@@ -86,11 +86,11 @@ function ConvertTo-Svg {
             }
 
             if ($WaterMarkText) {
-                if (Get-Command -Name 'New-WatermarkToSvg' -ErrorAction SilentlyContinue) {
+                if (Get-Command -Name 'Add-WatermarkToSvg' -ErrorAction SilentlyContinue) {
                     $opacity = [Math]::Clamp(($WaterMarkFontOpacity / 100), 0, 1)
-                    $null = New-WatermarkToSvg -SourceSvgFilePath $Document.FullName -WatermarkText $WaterMarkText -OutputSvgFilePath $Document.FullName -WatermarkTextFontColor $WaterMarkColor -WatermarkTextOpacity $opacity
+                    $null = Add-WatermarkToSvg -SourceSvgFilePath $Document.FullName -WatermarkText $WaterMarkText -OutputSvgFilePath $Document.FullName -WatermarkTextFontColor $WaterMarkColor -WatermarkTextOpacity $opacity
                 } else {
-                    Write-Verbose -Message 'WaterMark option requires New-WatermarkToSvg cmdlet from Diagrammer assembly.'
+                    Write-Verbose -Message 'WaterMark option requires Add-WatermarkToSvg cmdlet from Diagrammer assembly.'
                 }
             }